+++ tests/multimodal/test_gpu_ipc_memory.py
@@ -6,13 +6,41 @@ import time
 
 import pytest
 
+import vllm.config.multimodal as multimodal_config_module
+from vllm.config.multimodal import MultiModalConfig
 from vllm.multimodal.gpu_ipc_memory import (
     MultiModalGPUMemoryPool,
     get_mm_gpu_ipc_pool,
     maybe_init_mm_gpu_ipc_pool,
+    reserve_mm_ipc_gpu_memory,
     set_mm_gpu_ipc_pool,
 )
+from vllm.multimodal.video import (
+    PYNVVIDEOCODEC_CUDA_CONTEXT_BYTES,
+    PYNVVIDEOCODEC_DECODER_GPU_MEMORY_BYTES,
+    PYNVVIDEOCODEC_MAX_RETAINED_DECODERS,
+    PYNVVIDEOCODEC_VIDEO_BACKEND,
+)
 from vllm.utils.mem_constants import GiB_bytes
+
+
+def _mm_config(
+    *,
+    mm_ipc_gpu_memory_gb: float = 0,
+    video_backend: str | None = None,
+) -> MultiModalConfig:
+    video_kwargs = {} if video_backend is None else {"video_backend": video_backend}
+    return MultiModalConfig(
+        mm_ipc_gpu_memory_gb=mm_ipc_gpu_memory_gb,
+        media_io_kwargs={"video": video_kwargs} if video_kwargs else {},
+    )
+
+
+def _pynvvideocodec_decoder_budget(api_process_count: int = 1) -> int:
+    return api_process_count * (
+        PYNVVIDEOCODEC_DECODER_GPU_MEMORY_BYTES * PYNVVIDEOCODEC_MAX_RETAINED_DECODERS
+        + PYNVVIDEOCODEC_CUDA_CONTEXT_BYTES
+    )
 
 
 def test_acquire_release_accounting():
@@ -143,3 +171,72 @@ def test_global_pool_splits_budget_across_api_processes():
 def test_global_pool_rejects_invalid_api_process_count():
     with pytest.raises(ValueError):
         maybe_init_mm_gpu_ipc_pool(2, api_process_count=0)
+
+
+@pytest.mark.parametrize("video_backend", [None, "opencv"])
+def test_reserve_mm_ipc_gpu_memory_raw_frame_budget_only(
+    monkeypatch: pytest.MonkeyPatch,
+    video_backend: str | None,
+):
+    monkeypatch.setattr(
+        multimodal_config_module.envs,
+        "VLLM_VIDEO_LOADER_BACKEND",
+        "opencv",
+    )
+    mm_config = _mm_config(
+        mm_ipc_gpu_memory_gb=0.25,
+        video_backend=video_backend,
+    )
+
+    assert reserve_mm_ipc_gpu_memory(GiB_bytes, mm_config) == int(0.75 * GiB_bytes)
+
+
+def test_reserve_mm_ipc_gpu_memory_includes_pynvvideocodec_decoder_budget(
+    monkeypatch: pytest.MonkeyPatch,
+):
+    monkeypatch.setattr(
+        multimodal_config_module.envs,
+        "VLLM_VIDEO_LOADER_BACKEND",
+        "opencv",
+    )
+    mm_config = _mm_config(
+        mm_ipc_gpu_memory_gb=0.25,
+        video_backend=PYNVVIDEOCODEC_VIDEO_BACKEND,
+    )
+    available_bytes = 4 * GiB_bytes
+
+    assert reserve_mm_ipc_gpu_memory(available_bytes, mm_config) == (
+        available_bytes - int(0.25 * GiB_bytes) - _pynvvideocodec_decoder_budget()
+    )
+
+
+def test_reserve_mm_ipc_gpu_memory_uses_env_video_backend(
+    monkeypatch: pytest.MonkeyPatch,
+):
+    monkeypatch.setattr(
+        multimodal_config_module.envs,
+        "VLLM_VIDEO_LOADER_BACKEND",
+        PYNVVIDEOCODEC_VIDEO_BACKEND,
+    )
+    available_bytes = 4 * GiB_bytes
+
+    assert reserve_mm_ipc_gpu_memory(available_bytes, _mm_config()) == (
+        available_bytes - _pynvvideocodec_decoder_budget()
+    )
+
+
+def test_reserve_mm_ipc_gpu_memory_scales_decoder_budget_by_api_servers(
+    monkeypatch: pytest.MonkeyPatch,
+):
+    monkeypatch.setattr(
+        multimodal_config_module.envs,
+        "VLLM_VIDEO_LOADER_BACKEND",
+        PYNVVIDEOCODEC_VIDEO_BACKEND,
+    )
+    available_bytes = 8 * GiB_bytes
+
+    assert reserve_mm_ipc_gpu_memory(
+        available_bytes,
+        _mm_config(),
+        api_process_count=3,
+    ) == available_bytes - _pynvvideocodec_decoder_budget(api_process_count=3)
